+++ src/main/kotlin/dev/turingcomplete/intellijjvmsmanagerplugin/ui/detail/jvm/jvmaction/jtool/CreateJvmThreadDumpAction.kt
@@ -1,6 +1,8 @@
 package dev.turingcomplete.intellijjvmsmanagerplugin.ui.detail.jvm.jvmaction.jtool
 
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.project.Project
 import com.intellij.ui.components.JBCheckBox
 import com.intellij.unscramble.UnscrambleDialog
 import com.intellij.util.ui.UIUtil
@@ -32,7 +34,7 @@ class CreateJvmThreadDumpAction : JvmAction("Creating Thread Dump") {
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exposed Methods --------------------------------------------------------------------------------------------- //
 
-  override fun createComponent() = JPanel(GridBagLayout()).apply {
+  override fun createComponent(project: Project, parent: Disposable) = JPanel(GridBagLayout()).apply {
     border = UiUtils.EMPTY_BORDER
 
     val bag = UiUtils.createDefaultGridBag()