+++ src/main/kotlin/dev/turingcomplete/intellijjvmsmanagerplugin/ui/detail/jvm/jvmaction/jtool/GarbageCollectorAction.kt
@@ -1,7 +1,9 @@
 package dev.turingcomplete.intellijjvmsmanagerplugin.ui.detail.jvm.jvmaction.jtool
 
 import com.intellij.notification.NotificationType
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.project.Project
 import com.intellij.util.ui.UIUtil
 import dev.turingcomplete.intellijjvmsmanagerplugin.ui.action.RefreshProcessInformationAction
 import dev.turingcomplete.intellijjvmsmanagerplugin.ui.common.AnActionOptionButton
@@ -20,7 +22,7 @@ class GarbageCollectorAction : JvmAction("Garbage Collector") {
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exposed Methods --------------------------------------------------------------------------------------------- //
 
-  override fun createComponent() = JPanel(GridBagLayout()).apply {
+  override fun createComponent(project: Project, parent: Disposable) = JPanel(GridBagLayout()).apply {
     border = UiUtils.EMPTY_BORDER
 
     val bag = UiUtils.createDefaultGridBag()