+++ src/main/kotlin/dev/turingcomplete/intellijjvmsmanagerplugin/ui/detail/jvm/jvmaction/jtool/JvmMonitoringAction.kt
@@ -1,6 +1,8 @@
 package dev.turingcomplete.intellijjvmsmanagerplugin.ui.detail.jvm.jvmaction.jtool
 
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.project.Project
 import com.intellij.util.ui.UIUtil
 import dev.turingcomplete.intellijjvmsmanagerplugin.ui.common.AnActionOptionButton
 import dev.turingcomplete.intellijjvmsmanagerplugin.ui.common.UiUtils
@@ -17,7 +19,7 @@ class JvmMonitoringAction : JvmAction("Monitoring") {
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exposed Methods --------------------------------------------------------------------------------------------- //
 
-  override fun createComponent() = JPanel(GridBagLayout()).apply {
+  override fun createComponent(project: Project, parent: Disposable) = JPanel(GridBagLayout()).apply {
     border = UiUtils.EMPTY_BORDER
 
     val bag = UiUtils.createDefaultGridBag()