+++ src/main/kotlin/dev/turingcomplete/intellijjvmsmanagerplugin/ui/detail/jvm/jvmaction/jtool/ToStringJToolRunOption.kt
@@ -72,7 +72,6 @@ class ToStringJToolRunOption(optionTitle: String,
 
   // -- Private Methods --------------------------------------------------------------------------------------------- //
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
-  // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
   private class ShowProcessActionOutput(val taskTitle: String,
                                         val output: String,